Add listCell para as moedas

+++ src/database/CoinInfo.java
@@ -3,7 +3,7 @@ package database;
 public class CoinInfo {
     public String idCoin;
     public Double buyPrice;
-    public int qtd;
+    public Double qtd;
 
     @Override
     public String toString() {
